mention Notepad RS5 for #1
As of May 2018, Michel Lopez of Microsoft has announced that a forthcoming
update to Windows 10 will allow line endings other than CP/M (0D 0A), such as
the UNIX line endings used in zip or tar source packages.  This feature is
present in Insider and is expected to be released in the next Windows 10 service
pack codenamed "Redstone 5".
https://blogs.msdn.microsoft.com/commandline/2018/05/08/extended-eol-in-notepad/

However, even after RS5, users of Windows 7 or 8.1 are not expected to receive
the updated Notepad.  Thus issue #1 might remain open until January 2023, when
Microsoft plans to discontinue security updates for Windows 8.1.