lookup table for sprite hotspot X position
The character can use his abdominal and back muscles to flex his
spine to let his pelvis do more of the pushing.  My rotoscope actor
could not do so because of how he was folded, making the reference
footage I used to draw the movement somewhat unrepresentative.

So to give me room to play with making the character's spine
more flexible, change calculation of the frame 7 horizontal
offset to use a lookup table.